% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{BIC_lasso}
\alias{BIC_lasso}
\title{BIC Lasso}
\usage{
BIC_lasso(Y, betas, deltas, covariates, DIF_type, delta_unres, tolerance)
}
\arguments{
\item{Y}{data}

\item{betas}{vecotr of intercepts}

\item{deltas}{(ncov x nitems) matrix of DIF-parameters}

\item{covariates}{(n x ncovs) matrix of covariate values}

\item{DIF_type}{can be "items", "variables", "all.interactions", "none"}

\item{delta_unres}{DIF-parameter estimates from unrestricted estimation}

\item{tolerance}{Value under which parameters are considered equal
to zero for the calculation of degreees of freedom}
}
\value{
Returns the BIC for an estimated DIF model with (or without) penalization term
}
\description{
Calculates the BIC for an L1 penalized model. Needed for picking a
hyper parameter lambda.
}
