% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_itemvecs}
\alias{calc_itemvecs}
\title{Calculation Item parameters per person}
\usage{
calc_itemvecs(betas, delta, covariates)
}
\arguments{
\item{betas}{Itemparameter intercepts}

\item{delta}{DIF-parameter-matrix (nrow = ncovariates, ncol =  nitems)}

\item{covariates}{Matrix of covariate values of all persons (n x num. of covs)}
}
\value{
Matrix with itemparameters per person.
}
\description{
Calculates beta vectors per person based on covariate values.
}
