% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc_lambda_max_cml}
\alias{calc_lambda_max_cml}
\title{Calculate Lambda max for a sequence}
\usage{
calc_lambda_max_cml(beta_intercepts, covariates, Y)
}
\arguments{
\item{beta_intercepts}{beta_intercepts from restricted model, eg Rasch model without DIF}

\item{covariates}{Matrix of covariate values of all persons (n x num. of covs)}

\item{Y}{matrix of responses}
}
\value{
A list with the maximal Lambda-value for the sequence of Lambdas to be used for
optimization with three elements: selection for items, covariates and all interactions.
Note that for grouplasso this value is actually lambda_max * sqrt(groupsize), the groupsize
being either number of items or covariates, depending on the selection of penalization model.
The product is returned for these because group sizes are equal for all groups, so this
simplification is justified.
}
\description{
Calculates the maximum Lambda-value for the sequence (the minimum Lambda-value, for which all parameters are
equal to zero). For L1, this is the maximum-norm of the score function for the full modell evaluated at the
cml-estimates of the restricted model (hence all delta-parameters equal to zero). For Group-lasso this is the
maximum-norm of a vector containing the Euclidian norms per parameter group of the score vector evaluated
at the cml-estimates of the restricted model
}
