% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cl_person_2}
\alias{cl_person_2}
\title{conditional(log)Likelihoods}
\usage{
cl_person_2(x, betas)
}
\arguments{
\item{x}{0/1-vector of pattern of one person}

\item{betas}{vector of item easiness parameters, either for a person,
or a Matrix of calculated itemparameters for all persons already including
DIF-effects (npersons x nitems)}

\item{Y}{0/1-matrix of patterns of all persons (npersons x nitems)}
}
\value{
log conditional Likelihood for the person
}
\description{
Calculates the cLik/logcLik for one person or a whole sample given
parameters beta.
cl for one person via psychotools
}
