% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cl_sample_2}
\alias{cl_sample_2}
\title{conditional(log)Likelihoods}
\usage{
cl_sample_2(Y, betas, ...)
}
\arguments{
\item{Y}{0/1-matrix of patterns of all persons (npersons x nitems)}

\item{betas}{vector of item easiness parameters, either for a person,
or a Matrix of calculated itemparameters for all persons already including
DIF-effects (npersons x nitems)}

\item{x}{0/1-vector of pattern of one person}
}
\value{
logarithm of the conditional Likelihood for the sample
}
\description{
Calculates the cLik/logcLik for one person or a whole sample given
parameters beta.
logarithm of conditional Likelihood of the whole sample, adding them all together.
}
