% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cml_lasso_win_lbfgs}
\alias{cml_lasso_win_lbfgs}
\title{Title
cmlDIFlasso calculated with lbfgs}
\usage{
cml_lasso_win_lbfgs(
  data,
  covariates,
  DIF_type = "all.interactions",
  lambdas = 10^(-(-2:7)),
  tolerance_df = 0,
  ...
)
}
\arguments{
\item{data}{matrix of responses}

\item{covariates}{Matrix of covariate values of all persons (n x ncovs)}

\item{DIF_type}{can be "items", "variables", "all.interactions", "none"}

\item{lambdas}{vecotr of penalization parameters to be used}

\item{tolerance_df}{Value under which parameters are considered equal
to zero for the calculation of degreees of freedom}
}
\value{
list of lbfgs-optimization results for each lambda, the picked model
and its parameters, the computing time
}
\description{
Title
cmlDIFlasso calculated with lbfgs
}
