% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{cml_lasso_win_lbfgs_optimized}
\alias{cml_lasso_win_lbfgs_optimized}
\title{Title
Fitting cmlDIF with L1-penalty via lbfgs (opitmized with lambda_max)}
\usage{
cml_lasso_win_lbfgs_optimized(
  data,
  covariates,
  DIF_type = "all.interactions",
  nlambdas = 21,
  tolerance_df = 0,
  ...
)

cml_lasso_win_lbfgs_optimized(
  data,
  covariates,
  DIF_type = "all.interactions",
  nlambdas = 21,
  tolerance_df = 0,
  ...
)
}
\arguments{
\item{data}{matrix of responses}

\item{covariates}{Matrix of covariate values of all persons (n x ncovs)}

\item{DIF_type}{can be "items", "variables", "all.interactions", "none"}

\item{nlambdas}{number of lambda parameters evenly spread between 10^-5 and
lambda_max, which is calculated automatically}

\item{tolerance_df}{Value under which parameters are considered equal
to zero for the calculation of degreees of freedom}
}
\value{
list of lbfgs-optimization results for each lambda, the picked model
and its parameters, the computing time

list of lbfgs-optimization results for each lambda, the picked model
and its parameters, the computing time
}
\description{
Title
Fitting cmlDIF with L1-penalty via lbfgs (opitmized with lambda_max)

Title
Fitting cmlDIF with L1-penalty via lbfgs (opitmized with lambda_max)
}
