% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{expand_grid_JOHN}
\alias{expand_grid_JOHN}
\title{Faster verson of expand.grid for use in pairwise_prods}
\usage{
expand_grid_JOHN(seq1, seq2)
}
\arguments{
\item{seq1}{first vector}

\item{seq2}{second vector}
}
\value{
A matrix with two columns, the rows containing every possible combination
of the two vectors seq1 and seq2 ine element each.
}
\description{
Does the same as expand.grid in base R, just a lot faster.
Code is the fastest version taken from Stackoverflow:
https://stackoverflow.com/questions/10405637/use-outer-instead-of-expand-grid
}
