% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extract_obs_alpha_beta_power}
\alias{extract_obs_alpha_beta_power}
\title{Title
Helper function to extract information from result list with
wrap_cml_lasso_lbfgs_optimized results}
\usage{
extract_obs_alpha_beta_power(x, resultlist)
}
\arguments{
\item{x}{element for which to calculate}

\item{resultlist}{list of results from wrap_cml_lasso_lbfgs_optimized}
}
\value{
list of results and simulation scenario details.
}
\description{
Title
Helper function to extract information from result list with
wrap_cml_lasso_lbfgs_optimized results
}
