% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{loss_function}
\alias{loss_function}
\title{loss-function with penalization}
\usage{
loss_function(
  data,
  covariates,
  beta_intercept = rep(0, ncol(data)),
  delta_matrix = matrix(0, ncol(covariates), ncol(data)),
  DIFmodel = "RM",
  restr_pen = FALSE,
  lambda = 0.5,
  lambda_restr = 100
)
}
\arguments{
\item{data}{Answers of all persons (0/1, n x I)}

\item{covariates}{Matrix of covariate values of all persons (n x ncovs)}

\item{beta_intercept}{vecotr of parameter values for intercepts per item}

\item{delta_matrix}{Matrix (ncovs x nitems) of DIF-parameters per variable and item}

\item{DIFmodel}{can be "items", "variables", "all.interactions", "none".
And FALSE for Rasch model estimation}

\item{restr_pen}{For additional penalization to ensure zero sum of intercepts}

\item{lambda}{penalization parameter}

\item{lambda_restr}{penalization parameter for restriction}
}
\value{
the value of the (penalized) log conditional Likelihood
}
\description{
calculates logclik with or without penalization for DIF-model
given data, parameter values for items, DIF-parameters per variable per item
and value of lambda
}
