% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_df_from_resultlist}
\alias{make_df_from_resultlist}
\title{Title
Helper function to extract information from result list with
wrap_cml_lasso_lbfgs_optimized results}
\usage{
make_df_from_resultlist(resultlist)
}
\arguments{
\item{resultlist}{list of results from wrap_cml_lasso_lbfgs_optimized}
}
\value{
Dataframe with information per simulation result, including n, I, nvars, times,
observed alpha, power and others.
}
\description{
Title
Helper function to extract information from result list with
wrap_cml_lasso_lbfgs_optimized results
}
