% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pairwise_prods}
\alias{pairwise_prods}
\title{Pairwise products of two vectors}
\usage{
pairwise_prods(vec1, vec2)
}
\arguments{
\item{vec1}{first vector}

\item{vec2}{second vector}
}
\value{
Matrix with all products (length(vec1) = nrows, length(vec2) = ncols))
}
\description{
Given vectors 1 and 2, delivers matrix with pairwise products.
Necessary for calculation of the score function.
}
