% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{penalization}
\alias{penalization}
\title{penalization term}
\usage{
penalization(
  deltas,
  beta_intercept,
  DIF_type = c("items", "variables", "all.interactions", "none"),
  restr_pen = FALSE,
  lambda = 0,
  lambda_restr = 0
)
}
\arguments{
\item{deltas}{A (ncovars x nitems) matrix of DIF-parameters}

\item{beta_intercept}{A vector of #items parameters}

\item{DIF_type}{can be "items", "variables", "all.interactions", "none"}

\item{restr_pen}{For additional penalization to ensure zero sum of intercepts}

\item{lambda}{penalization parameter}

\item{lambda_restr}{penalization parameter for restriction}
}
\value{
A numeric value for penalization of the Loglikelihood
}
\description{
Calculates the certain amount of penalization depeding on DIF-type,
lambda and for restriction of intercepts added
}
