% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RM_can_nlm}
\alias{RM_can_nlm}
\title{Rasch Model optimization with own loss}
\usage{
RM_can_nlm(starting_values, data, restr_pen = FALSE, optimization = "nlm")
}
\arguments{
\item{starting_values}{starting values for optimization}

\item{data}{matrix of response data}

\item{restr_pen}{To conduct penalization for sum=0 restriction set this to TRUE.}

\item{optimization}{either "optim" or "nlm"}
}
\value{

}
\description{
RM_optimization via loss_function and cl_sample_2, no penalization.
}
