% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RM_DIF_lbfgs}
\alias{RM_DIF_lbfgs}
\title{Rasch model with DIF via lbfgs (only L1)}
\usage{
RM_DIF_lbfgs(
  data,
  covariates,
  starting_betas = rep(0, ncol(data)),
  starting_deltas = matrix(rep(0, ncol(data) * ncol(covariates)), ncol(covariates),
    ncol(data)),
  lambda = 0,
  ...
)
}
\arguments{
\item{data}{matrix of responses}

\item{covariates}{Matrix of covariate values of all persons (n x num. of covs)}

\item{starting_betas}{beta_intercepts starting values for all items (vector)}

\item{starting_deltas}{starting values for DIF-parameters as matrix (ncovs x nitems)}

\item{lambda}{penalization parameter}

\item{...}{other arguments to be passed to lbfgs}
}
\value{
result of optim-optimization for DIF-model via cML-loss-function for
a certain penalisation-type (for now only all.interactions) and lambda value
}
\description{
optimizes loss_function with DIF-part for certain tuning-parameter
lambda with lbfgs
}
