% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{RM_DIF_nlm}
\alias{RM_DIF_nlm}
\title{Rasch model with DIF via nlm}
\usage{
RM_DIF_nlm(
  data,
  covariates,
  DIF_type,
  lambda,
  starting_betas,
  starting_deltas,
  restr_pen = TRUE,
  lambda_restr = 100,
  ...
)
}
\arguments{
\item{data}{matrix of responses}

\item{covariates}{Matrix of covariate values of all persons (n x num. of covs)}

\item{DIF_type}{should DIF be calculated? options: "items", "variables", "all.interactions"}

\item{lambda}{strength of penalization}

\item{starting_betas}{beta_intercepts starting values for all items (vector)}

\item{starting_deltas}{starting values for DIF-parameters as matrix (ncovs x nitems)}

\item{restr_pen}{For additional penalization to ensure zero sum of intercepts}

\item{lambda_restr}{penalization parameter for restriction}

\item{...}{other arguments to be passed}
}
\value{
result of nlm-optimization for DIF-model via cML-loss-function for
a certain penalisation-type and lambda value
}
\description{
optimizes loss_function with DIF-part for certain tuning-parameter lambda
}
