% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{score_func_person}
\alias{score_func_person}
\title{Score function for one person}
\usage{
score_func_person(beta_person, y_person, covariates_person)
}
\arguments{
\item{beta_person}{Parameter vector of a certain person.}

\item{y_person}{Pattern of answers of a person.}

\item{covariates_person}{Covariate values of a person.}
}
\value{
a list containing the values of the score function
\item{scoref_betas}{Values of the score function for every intercept.}
\item{scoref_delta}{Values of the score function for every delta-parameter (ncovs X nitems)}
}
\description{
Calculates the derivatives of the conditional Loglikelihood for all
intercept and delta parameters.
}
