% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{score_func_sample}
\alias{score_func_sample}
\title{Score values of the sample per parameter}
\usage{
score_func_sample(beta_intercepts, delta, covariates, Y)
}
\arguments{
\item{beta_intercepts}{Vecotr of beta intercept values.}

\item{delta}{A (ncovs x nitems) matrix containing the DIF-parameter values.}

\item{covariates}{A (npersons x ncovs) matrix containing the covariate values.}

\item{Y}{(npersons x nitems) matrix of answers of all persons to all items.}
}
\value{
a list containing the values of the score function
\item{score_betas}{Values of the score function for every intercept.}
\item{score_deltas}{Values of the score function for every delta-parameter (ncovs x nitems)}
}
\description{
Calculates the derivatives of the conditional Loglikelihood with respect
to each intercept and delta parameter for the full sample.
}
