% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{session_all}
\alias{session_all}
\title{Title
Session information for simulation and estimation}
\usage{
session_all(no_of_cores_used = "unknown")
}
\arguments{
\item{no_of_cores_used}{number of cores used for anlysis.
To be completely correct, this should be the number of threads used.}
}
\value{
A list including the output of sessionInfo(),
the CPU name, the amount of system RAM,
the number of cores used (to be completely correct, this is the number of threads used),
the number of threads available and the number of physical cores available
}
\description{
Title
Session information for simulation and estimation
}
