% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sim_scenario_DIF_data}
\alias{sim_scenario_DIF_data}
\title{Simulating DIF-data outputting scenario}
\usage{
sim_scenario_DIF_data(
  npersons,
  nitems,
  ncovs,
  seed,
  betavalues = rep(0, nitems),
  deltavalues = matrix(c(-0.5, 0, 0, 0, 0.5), ncovs, nitems),
  thetavalues = rep(0, npersons),
  covariatevalues = matrix(rnorm(npersons * ncovs), npersons, ncovs)
)
}
\arguments{
\item{npersons}{number of persons}

\item{nitems}{number of items}

\item{ncovs}{number of covariates}

\item{seed}{seed to be set for simulation}

\item{betavalues}{beta_intercepts for all items (vector)}

\item{deltavalues}{values for DIF-parameters as matrix per item (cols) per covariate (rows)}

\item{thetavalues}{vector of person parameters}

\item{covariatevalues}{Matrix of covariate values of all persons (n x ncovs)}
}
\value{
randomly drawn response matrix for DIF-model and given parameters
}
\description{
Draws binary data randomly for given item intercepts, DIF-parameters,
covariate values and person parameters. Returns a list with the simulated
data and all the parameters used to draw the sample.
}
