% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_DIF_data}
\alias{simulate_DIF_data}
\title{Simulating DIF-data}
\usage{
simulate_DIF_data(betas, delta, covariates, thetas = 0)
}
\arguments{
\item{betas}{beta_intercepts for all items (vector)}

\item{delta}{values for DIF-parameters as matrix per item (cols) per covariate (rows)}

\item{covariates}{Matrix of covariate values of all persons (n x ncovs)}

\item{thetas}{vector of person parameters}
}
\value{
randomly drawn response matrix for DIF-model and given parameters
}
\description{
Draws binary data randomly for given item intercepts, DIF-parameters,
covariate values and person parameters. (Based on easiness parameters.)
}
